Fix selection logic in Software spoke (#1404158)
Removed selectedGroups which is not helpful and use addonStates
dictionary instead. The addonStates variable has 3 states, SELECTED,
DESELECTED and DEFAULT. When you change the state to the environment
default, it will return state in the addonStates variable to DEFAULT.

Remove code for removing groups from selectedGroups array when changing
environment; this is not required anymore when the state is read from
buttons.

Groups marked for installation are read from GTK now.

Resolves: rhbz#1404158

(based on commit 12299ab2e0a1c1f0781808fe791d1dfaafb36117)

Reported-by: Vit Ry <Frodox@zoho.com>