yum: Always release payload and catch raised exceptions.
Payload should be released also in a case that a dependency error is
found.

Release can raise an exception if we try to release resources from
different threads. That should not happen and we should fix these
cases. However, we don't want the installer to fail because of that
either. Therefore, all exceptions from release will be catched and
logged only.

Related: rhbz#979307