Support --nohome option in the autopartitioning (#663099)
Default partitioning can be modified in an install class's method
customizeDefaultPartitioning, where can be applied selections made in
a kickstart file. This method can be easily extended for future
--noswap and --noboot options.

The base install class will remove a request for a /home partition,
if --nohome option is selected in a kickstart file. In that case,
the /home partition will not be created.

Subclasses of the base install class can easily ignore this
customization if they override the customizeDefaultPartitioning
method with the pass statement.

Resolves: rhbz#663099