Fix partial kickstart software selection in GUI (#1404158)
When groups are selected in %packages section in a kickstart file, mark
these groups in the Software spoke. However if the user visits the
Software spoke the settings from kickstart (environment & groups) will
be replaced by what the user selects in the spoke. Individual package
selection from kickstart should be unaffected.

Kickstart file without user interaction is the only option how to
install groups and packages which are not inside of the selected
environment.

Related: rhbz#1404158