Fix partial kickstart software selection in GUI (#1404158)
When groups are selected in %packages section in a kickstart file, mark
these groups in the Software spoke. However if user visits the Software
spoke it will erase kickstart settings and use user settings.

Kickstart file without user interaction is the only option how to
install groups and packages which are not inside of the selected
environment.

Related: rhbz#1404158

(based on commit e52d17b320cc433443372d6a8275928dba8b8776)