Enhance git-find-branch script
The git-find-branch script worked badly if we were on a main branch
(eg. f26-devel) and this branch was pointing to the same place as
master branch (or some other branch with higher priority).

This is fixed now by adding quick test which is testing branch name.
If this name is the same as some main branch it will do another check to
test if they are pointing to the same commit.

This solution is also quicker when you are on a main branch directly
or for CI.

NOTE:
You can still break this by changing name of your branch or creating new
one from the base which doesn't have any new commit.