gui: show supported locales on Atomic Host installs
Fedora Atomic Host has never shipped with all supported locales. It
originally shipped with only en_US, though that was changed afterwards
to cover at least the same locales as RHEL7[1]. However, we would still
show all the locales in the installer GUI.

This patch adds support for allowing installclasses to filter locales
that should be shown based on the actual content carried in the ISO. The
Atomic installclass can then make use of this. This is essentially the
same patch as the one for rhel7-branch[2], but split out since the
Atomic installclass is not here.

[1] https://pagure.io/fedora-atomic/c/5a5edddc9a6b943e6fe902defad177009c43867a
[2] https://github.com/rhinstaller/anaconda/pull/871

Related: https://pagure.io/atomic-wg/issue/282