Display progress for the post installation phase
Instead of just telling the user that we are "Performing
post installation setup tasks" for many minutes display
detailed progress for the post installation phase.

Both post-scriptlets and installed package verification
are now displayed in the UI as well as logged in packaging.log.

Scriptlets that are run during the installation phase are just logged,
but not displayed in the UI to not needlessly interrupt the flow of
"installing foo-bar 15/345" messages.

Also raise the required DNF version to one providing the corresponding API.