Make geolocation with kickstart possible (#1358331)
Make it possible to use geolocation during a kickstart based
installation. Geolocation is disabled by default during a kickstart
installation to make it more deterministic and independent on geographic
location of the machine.

To enable geolocation during a kickstart installation, use the
inst.geoloc-use-with-ks=1 option (or the --geoloc-use-with-ks=1 option
when running Anaconda as an application from the command line).

There is also an API that install classes can use to enable geolocation
during kickstart installations - the use_geolocation_with_kickstart
attribute.

Resolves: rhbz#1358331