Call subprocess.Popen with absolute path to a binary (#1411407)
Resolves: rhbz#1411407

Python is trying to execve the program (specified without the path) in the
fork, iterating over the path list and calling strerror when failing on paths
before it finds the right one. The strerror call can then hang waiting forever
for a lock held by parent process (__libc_setlocale_lock).

NOTE: in the future we want to replace this workaround by using subprocess32
module.