Tweak epoch definition to fix system clock setting (#1433560)
This approach was proposed by rmarshall, I am just submitting
it as he didn't get around to it yet. I have tested this and
verified it works: the correct time is calculated and set even
when Python's concept of the 'current' timezone isn't the same
as the one the function is called with.