Trigger the entered signal only once the screen is shown (#1443011)
Trigger the entered signal only once the screen corresponding to the
given spoke or hub is shown.

This should better correspond to the signal name (we only trigger it once
the screen is visibly entered) and also makes it possible for spokes to set
(and unset) their own custom key bindings.

Also thanks to Vojta Trefny and Vendula Poncova for helping me to figure
out the correct method calling order. :)