Add AnacondaPrefixFilter
Add AnacondaPrefixFilter - basically a pass-through filter (never
discards any log records), that sets the log_prefix attribute for
log records.

The logic is simple: - if the name attribute starts with "anaconda.",
it drops the "anaconda." prefix and sets log_prefix to the remainder.

So for example name "anaconda.geoloc" gets a log prefix "geoloc"
and "anaconda.payload.rpmostree" gets a log prefix "payload.rpmostree".

If the log record name is just "anaconda" the record is considered to
be something targeting Anaconda in general, so the log_prefix
attribute os set to "misc" for miscellaneous.