Really fix with tmux 2.4 (version comparison was busted)
The attempt in 6752b131 to conditionalize the tmux config on
tmux version didn't work right, for two reasons: bc isn't in
the installer environment, and bc doesn't do the right kind of
comparison anyway (we want to compare following versioning
conventions, bc does a floating point number comparison). This
ought to do it properly, I hope.