Show warnings about enabling the network time (#1418604)
When the network time cannot be enabled or disabled, we show a
warning message and reset the switch button to the previous state.
With the reset, the handler is run for the second time and clears
are warning messages, so the user doesn't see the message at all.
Unfortunately, we cannot use handler_block_by_func, because it does
not work, so we need to block the handler manually.

Resolves: rhbz#1418604