Make it possible to use a custom getpass()
Make it possible to use a custom getpass() implementation.
The getpass() function provided by the Python standard library
is hardcoded to always expect input from /dev/tty.
This is unfortunately not enough for the Initial Setup multi-tty
usecase, where password input has to be supported from arbitrary
consoles. So make it possible to use a custom getpass() implementation
that does not have this limitation and can run on an arbitrary console.