Set the default filesystem type from a kickstart file
If the kickstart file specifies the --fstype option in the autopart
command, we should use it to set the default filesystem type. If the
type is not set in a kickstart file, we will use the specified type
from an install class or the default type from blivet.

Also, done a little cleanup in the storage property.