Improve the GUI password checker
Provide a mechanism for detecting if the content of the input and confirmation
fields has changed since the screen when password checking is taking
place has been displayed to the user.

Also properly provide password quality check feedback when overriding
password/passphrase set in kickstart & the policy allows that (--changesok).

Related: rhbz#1356966