Use unique 3 letter log level names
Assign unique 3 letter log level names to improve
readability of the log files.

Previously the log level names had variable length,
making the log files less readable due to log messages
beginning on a variable offset on lines in the log file.

With the fixed with name log messages start all at the same
offset from the start of the line and the lines might also be slightly shorter,
so more text might fit on the screen when viewing the logs during installation,
especially on low resolution screens.