Delete file systems used only by a given installation (#1453097)
If user chooses to delete all file systems which are only used by
a given installation, we will skip the file systems that are shared
between more installations.

Before that we offered to delete all systems including those
used by other installed operating systems. That could be confusing
and result in data loss. The new behaviour is closer to Fedora's.

Resolves: rhbz#1453097