Fix a typo in python-meh initialization (#1462825)
Fix a typo in initialization of the python-meh provided Config class.

A while ago the fileList parameter was renamed to file_list on Anaconda
side due to a small refactoring error. Unfortunately the __init__()
method of the Config class silently accepts *any parameter names whatsoever*,
even if they don't match the parameters it actually accepts.

End result: The additional log files Anaconda specifies are not included
            in the dump as file_list != fileList.

So let's:
- fix parameter name
- add "Make the __init__() method of the python-meh Config class sane"
  to our TODO list