rpmostreepayload: Reuse the local repo as a cache
When doing kickstart installs (can happen for both bare metal PXE as well as
building cloud images), we currently end up redownloading all of the ostree
content, even though we have a version in the squashfs already.

However, the version we have may not be the version we want; the installer image
is typically built at a different/slower cycle from the ostree content.

But this is where a strength of libostree's content-addressed storage shines; we
can use the local repo as an object cache, and avoid fetching what we already
have.

This just enables the logic that lives in libostree.

Closes: https://github.com/rhinstaller/anaconda/issues/1117