rpmostreepayload: Explicitly create /var/lib before tmpfiles
Currently ostree hardcodes creating e.g. `/var/lib`, see:
https://github.com/ostreedev/ostree/blob/82024f161b280b4ef70880ddd5bc83ad812e6f0a/src/libostree/ostree-sysroot.c#L1388

I'd like to remove that, as it doesn't work in the case where in Anaconda the
admin sets up a `/var` mountpoint. Unfortunately, we can't just tell
`systemd-tmpfiles` to use `--prefix=/var/lib`, as that will create a lot of
directories for which we don't have NSS configuration, and there's no
`--only=/var/lib` or similar.  Hence we just `mkdir -p` it explicitly.