Add support for automatic generating of DBus specification.
With provided generators dbus_signal, dbus_interface and dbus_class,
it is possible to define a DBus interface with a Python class and
provide a DBus specification of a Python class. Types are derived
from type hints using types from pyanaconda.dbus.typing. Only class
members with CamelCase names are considered to be members of the DBus
interface.

It is supported the inheritance of DBus interfaces. Generated DBus
specification will contain all interfaces of the class ancestors.
If the class defines a new DBus interface, then class members with
names already defined in inherited interfaces are not considered to
be members of the new DBus interface.

Example of a class that defines a new DBus interface with a method,
a signal and a property. Instances of this class can be also exported
to DBus.

@dbus_interface("Interface")
class NewInterface(object):

  def Method(self, x: Int) -> List[Int]
    return [x]

  @dbus_signal
  def Signal(self, y: Double)
    pass

  @property
  def Property(self) -> Bool:
    return False

Example of a class that doesn't define a new DBus interface, but
instances can be exported to DBus.

@dbus_class
class ExportableClass(NewInterface):
  pass

It is required DBus library pydbus.