rpmostreepayload: Rollup backport of Fedora patches
This rolls up all of <https://github.com/rhinstaller/anaconda/pull/1129> which
is basically mount handling, plus the pull-local optimization for kickstarts.

We need to change the implementation for the mount setup, since
things work differently in RHEL7 vs Fedora since commit
664ef7b43f9102aa9332d0db5b7d13f8ece436f0.

The code that changed in `pyanaconda/install.py` only applied to ostree-based
payloads (where we introduced the physical vs sysroot distinction), so should be
safe for backporting.

Resolves: rhbz#1459623