dnfpayload: do not try to contact disabled repo
This patch fixes a regression introduced by:

79e66d2a6 Fix Anaconda forces payload restart when network (not)change (#1373449)

When anaconda tries to build the repositories MD5 cache, it mistakenly
iterates over all the repositories it is aware about, ever the disabled ones.

The end result is that it tries to reach the default disabled repositories from
"/etc/anaconda.repos.d/", for example
"https://codecs.fedoraproject.org/openh264/$releasever/$basearch/".

If the user is in a private intranet and the internet is not reachable, this
may lead to very long delays related to long dnf timeouts.

The fix is to only build the MD5 cache for the enabled repositories and leave
the disabled ones alone.

Resolves: rhbz#1470552

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>