Network spoke freeze when testing availability (#1478970)
The Network spoke was blocked for some time when the repo wasn't
reachable. In case there was a connection issue when user clicked
on the Done button it froze for a 5 minutes (default timeout).

The time of the freeze was reduced in previous patch however it still
froze for a few seconds. To avoid this problem, the repo availability
testing have been moved to payload thread now.

Information about test of existing repositories is now printed as
status message in source spoke.

Related: rhbz#1478970