Add Dialog and PasswordDialog TUI objects
This is the replacement for the EditTUIDialog and EditTUISpoke objects
functionality.

Both classes are designed to be used without an inheritance. This helps
to have less intertwined code and also easy to use in an existing places.

Both classes are using new Simpleline solution which is asking for
user input in separate loop as a modal window. This should prevent more
issues when multiple screens wants input in parallel.

Add decorators for easier work with checks and reporting.