Fix the race between StorageCheckHandler and updateCompleteness in main hub
This happens approximately in 1 of 100 deployment with next conditions:

- extremely slow storage IO
- kickstart installation with slightly reduced min_partition_size which causes
  appropriate warning

The StorageCheckHandler reports than StorageSpoke is ready, then start
logger for the warning. In separate thread hubQ got HUB_CODE_READY,
but StorageSpoke is not completed because of THREAD_CHECK_STORAGE is not
finished. So, installation stops on main Hub with the warning.