Use build-time release (#1493952)
As we shown the Anaconda version including the
release we really should use the release as it
exists at the package build time instead of
the release at tarball creation time (which
is generally always 1).

Otherwise we would be lying to users every time
the Anaconda package has been rebuild without
changing the tarball (patch only rebuilds,
mass rebuilds, modular builds, etc.).

So ad a new configure variable that can be used
from the spec file at build time and use it
to set the built-time release version.

NOTE: The release version might not be a number,
but basically an arbitrary string as with
the modular builds.

Resolves: rhbz#1493952