Add changelog entries from the unstable branch
We have decided to start doing releases directly from the master
branch again. Like this it's easier for people to use latest
Anaconda from git, while still having information what goes to
which release. The unstable branch only got updated right before
release and was not tracking the master branch continuously.

This also fixes the confusing issue of having an ancient release
number set in the master branch due to release bumps only happening
on the unstable branch.

So add the spec file changelog entries from the unstable branch to the specfile
in the master branch to avoid loosing changelog continuity.