Do not execute storage when the spoke is left with no selected disk (#1496327)
Leaving the storage spoke with all disk unselected is currently there as a way
to get back out of the spoke. In this special case the storage is not reset (to
snapshot) when leaving the spoke, so it doesn't make sense to try to re-execute
the storage on not updated (reset) device tree as we do. It would lead to
invalid results (not enough space on the unselected disks which are actually
used for execute/partitioning).

Could resetting the storage/devicetree also in this special case help? It would
not work because no selected disk in GUI just results in setting storage's
exclusive_disks to empty list (with ignored_disks being kept empty list) which
actually does not hide any disks in storage reset (leading to the invalid result
described above).

So let's just leave the spoke while ensuring it is in incomplete state (by
setting the checker error to "No disks selected").