Fix dnf exception repository not set (#1495211)
The set_source_url() is already called in SpecifyRepoSpoke. It shouldn't
be called before because user may decide to not set anything as a repo.

To fix problem with changing installation method the ksdata.url is used
instead of the ksdata.method which is pointer to the active installation
method.