Add support for logging from DBUS modules
The plan is that modules will always log to the Journal
and any aggregated log will be generated from filtering
the Journal message stream.

For that the Journal setup function is needed outside
of the main logging class, so let's move it to module top level.

Next we add a DBUS module logger getter, which does two things:
- sets up logging - each DBUS module is in it's own process,
  so it needs to make sure it's logging is properly setup
- returns the corresponding logger

Once this method is called, normal Anaconda get_module_logger()
calls will work as expected.