Add .service and .conf files for the DBUS modules
How the DBUS modules (including Boss) started ?

The Exec= directive contains:

/usr/bin/python3 -m pyanaconda.modules.<module_name>

where <module_name> is the name of the module (basically it's folder name)
to be started. Once this happens, the __main__.py file is run by the
Python interpreter. The __main__.py file is responsible for correctly
starting the DBUS module, which includes publishing it's interface on
DBUS as well as starting its mainloop.

Using python -m for starting DBUS modules has a range of benefits:
- no need to place executables separately from the Anaconda and DBUS module code
- simplified updates image creation and general debugging due to things
  being in one place
- no need to change the path in Exec= in the service files as PYTHON_PATH
  manipulation can be used instead to easily start DBUS modules from a local
  checkout instead of from a system-wide locations
- no need to store many executable files in the Anaconda source tree or
  to make them executable during build
- __main__ only does a very limited amount of work and all other modules are
  always imported, so they can use the __name__ based logger getters just fine
- all code for an individual DBUS module is in a single folder, so
  relative import can be easily used to access DBUS module code present
  in individual files in it's folder