Add a script for running Boss & modules locally
The script uses GTestDBus to start a custom DBUS daemon
and will then run Boss and all the modules on it's bus.

This makes it fairly straightforward to debug Boss and DBUS modules
directly from the Anaconda source checkout and without the need for
root privileges.

Also by using a custom DBUS daemon instance there is no risk of
breaking you actual session bus services and as a side effect
each run is totally separate from the previous one.