Fix changing source don't erase old environment TUI (#1505090)
When the installation source changed the environment in the software
selection TUI spoke stays the same. Because of this when user changed
source and leave the same settings for the environment and addons as
before then the payload wasn't restarted at all.

Fixing by removing old saved environment after payload restart.

Resolves: rhbz#1505090