Refactor DASD formatting and support detection of LDL DASDs.
The code for DASD formatting was moved to the class DasdFormatting in
the pyanaconda.format_dasd module. Automatic formatting can be run
with run_automatically, otherwise UI has to do the steps manually.
This way the logic is separated from the UI and it is the same for
TUI and GUI. And it will be easier to move the code on DBus.

Also, formatting of LDL DASDs is now supported as well. Disks with the
LDL format will be automatically formatted with the clearpart --cdl
option. Otherwise, user can choose to format them if he selects them
in a spoke.

Based on commits:
a4a2e86, 7e28f03, aab9b91, 356e89c, d83e06d, 4ffe0de, 1e0f384