Really install all the right packages on Mac UEFI installs
My previous fix still wasn't quite right - it overrode the
EFIGRUB `_packages64` list so those packages, which are needed,
weren't included. So let's *add* to that list, not replace it.
As it's not possible to do this with a class variable, let's
make `_packages64` an instance variable in all cases, for
consistency.