Add a unit file for Boss startup
To start Boss & prototype module use inst.run_boss on the
boot command line as Boss is not yet started by default.

Also make sure the file is included in the source tarball
and the corresponding Systemd service is started in
the installation environment (eq. is part of the Anaconda target).