Replace get_bus with the class DBus
We should call static methods of DBus class instead of get_bus. The
class DBus keeps one instance of the DBus connection that is created
on demand. It also provides methods for easy publishing and creating
proxies.

We should never keep our own instances of the DBus connection unless
we have a very good reason for it.