Tweak run_boss_locally script
The script can be started from anywhere now.
The bus address will be printed and wait for user to copy it and connect
to the buss.
The bus address will be printed green if python3-colorama package is
present, otherwise the warning about missing package will be printed but
the script should work as usual.