Modules should use the proxy pattern.
The class that describes the interface of a kickstart module should
inherit the KickstartModuleInterface class and the class that provides
the implementation of the interface should inherit the KickstarModule
class.

Kickstart modules can define KickstartSpecification to be able to
read and write a kickstart string with expected commands and sections.