Improve password checking status and error messages
Most importantly have two variants, one for passwords and one for
passphrases. Related to that use "secret" for stuff that can be
password or passphrase (and possibly also something else in the future).

Instead of "password name" a secret-type enum is now used to
differentiate if the secret being checked should be called a
password or a passphrase.

And when we are at it make "Done" bold in the warning messages,
to hopefully improve their readability.