Variants need to be instances of the Variant class
When we send on DBus values of the type Variant, these values need to
be packed into instances of the Variant class. Use get_variant to get
a new instance with a type defined by a type hint.

For example:
v = get_variant(Bool, True)

It is the same as:
v = Variant('b', True)

Also ValueErrors are changed to TypeErrors and the mypy test was
removed, because it wasn't possible to run it anymore.