Add an object observer with cached properties
Class DBusCachedObserver is an extension of DBusObjectObserver that
allows to cache properties of the remote object. The observer is
connected to the PropertiesChanged signal of the remote object and
updates its cache every time the signal is emitted. Then it emits its
own signal cached_properties_changed. Only properties of specified
interfaces will be cached.