Connect to the observed service and other stuff.
Call connect_once_available to wait for a signal before we can access
objects of the observed service. Call connect to be able to access
the objects immediately.

Also, the cached observer can be forced to update its cache with
force_update method.

Also, messages in raised exceptions are fixed.