UI support for configuring certain mitigations (#1534833)
Add the "KERNEL & PERFORMANCE" spoke to GUI and TUI, which makes
it possible to turn off certain security mitigations.

Mitigations that can currently be turned off:

- PTI
- IBRS
- IBPB

The mechanism is similar to (and compatible with) using the
--append option of the bootloader kickstart command.

If a mitigation is turned off via bootloader --append
(for example with "bootloader --append 'nopti') this
will be reflected in the UI. And turning mitigation on and
off will be reflected in the output kickstart.

Resolves: rhbz#1534833