Remove false positive but disable Pylint in makeupdates script
The python3 koji library is used only in makeupdates script but Pylint
will raise error because of this import in build requires. I guess we
don't want to add another dependency which is not used in Anaconda and
not part of the package.

On the other hand we don't want to have it ignored globally.