Add support for different message buses
Class Connection is an abstract class that provides a functionality
for working with a bus connection. The connection should be created in
the abstract method get_new_connection. Call methods get_observer and
get_cached_observer to get observers of a remote object.

Classes DBusSystemConnection and DBusSessionConnection define
connections for the system and session buses. DBusConnection define
a connection to a bus at the specific address. DBusDefaultConnection
defines a connection that should be used by all Anaconda modules
to communicate with each other.

Objects DBus, SystemBus and SessionBus provides the connections to
the default, system and session buses.

The class DBusServiceObserver was removed, because it was too complex.