Fix mitigation toggling with kickstart
It turns out that we can't depend for the execute()
method of the bootloader command to be executed
at the start of the installation phase. Whilte it's
called at start of a fully interactive installation,
it's called very early on (at spoke initialization)
during partial kickstart installation and then not
called again.

So instead, make sure the apply the mitigation options
eac time user exists the spoke via the spoke apply() method.

Related: rhbz#1534833