Only show Kernel & performance spoke on x86
The kernel and performance spoke provides options to turn
of the following mitigations: KPI, IBRS and IBPB.

Unfortunately it turns out these options are not generally available
on non-x86 RHEL 7 architectures.

So only show the Kernel and performance spoke on x86 and hide it on
all other architectures.

Related: rhbz#1534833