Move system-logos dependency from anaconda-core to anaconda-gui
The system-logos package is needed in two places:
- by the Anaconda GUI
- by installation-environment-only non-GUI tools such as
  grub2 or syslinux[0]

Making it an anaconda-core dependency results in it being dragged in
also when it is not needed, such as when Anaconda with TUI is installed as an
application or as Initial Setup TUI dependency.

So add system-logos to the corresponding Lorax template[1] and move the
system-logos dependency from anaconda-core to anaconda-gui.

[0] https://github.com/rhinstaller/anaconda/pull/1289
[1] https://github.com/rhinstaller/lorax/pull/327

Related: rhbz#1529239