Migrate Anaconda to our private dbus session (#1551096)
Configurations are placed in /usr/share/anaconda/dbus/confs
DBus service files are placed in /usr/share/ananconda/dbus/services
And main configuration file is placed in /usr/share/anaconda/dbus

Thanks to this change we can setup permission as we need and change them
when we need. Also it will avoid situations when user connects to
a module and change its settings in an unexpected way.
As a bonus for debugging we can easily trace traffic on our DBus without
interference of other services.

Related: rhbz#1551096