Initial module enablement and installation support
Use new DNF API entry points to make it possible to enable and install
modules. At the same time we can radically simplify our group exclusion
and environment resolution code by making full use of the new API.

The main change is that instead of incrementally calling "install package"
and "select group" we now gather all the things we want to include
and exclude in the installation transaction and feed it to DNF at once.

Also bump minimal DNF version to make sure the new entry point is
available.