Add anaconda-install-env-deps as dependency of the anaconda package
This way anyone installing just the "anaconda" package should still get
all the install time dependencies as before (seems to be the case at
least with Pungi).

And at the same time. Initial Setup should be fine, as it depends only
on "anaconda-tui" or "anaconda-gui", which both depend on "anaconda-core",
but not in any way on the "anaconda" package itself.