Support temporary kickstart generating
Sometimes we extend a DBus module with new kickstart commands, but
the kickstart commands are still handled by UI for a while. Then
we need to be able to generate a temporary kickstart representation,
that doesn't contain these unsupported commands, because they will
be generated by UI. This functionality is now provided by the method
GenerateTemporaryKickstart.