Select Workstation install class for Workstation live (#1569083)
The Workstation install class (thus default settings, branding)
is not currently selected for Workstation live image installs,
ever since the install class was moved into anaconda itself.
The most correct fix for this would likely be for lmc to include
.builddstamp in live images, then the existing detection code
should work, but as a quick fix for F28, let's parse the variant
ID out of /etc/os-release.

Resolves: rhbz#1569083

Signed-off-by: Adam Williamson <awilliam@redhat.com>