Handle empty active attribute for consoles (#1569045)
Some consoles are not real consoles, but fake/alias consoles
that point to the real console via the "active" attribute.
Apparently it is also possible to have a chain of fake/alias
consoles pointing to each other where only the last one
points to a real console (one that has no active attribute).

But for some reason on some PPC systems there are consoles that
have the active attribute, but it is empty (rhbz#1569045).

If we hit this situation, we can't continue our search for the
real console as we don't know where to go next. So log a warning,
set the console name to "" and stop the search.

This way the code will no longer traceback, but it might
be unable to resolve the correct active as long as the chain
if fake/alias consoles is broken by containing the empty active
attribute.

Resolves: rhbz#1569045